Updated HackerForm
Fixed some issues, as shirt size null in form.
Added a checkbox in the last step to set true if accept terms and condicions.
Submit button would be disabled if checkbox is not checked.
Fix notification value, before was value.acceptNotifications, now is value.notifications.
Set value of notifications in body>config.